# api/swagger/swagger.json
{
  "info": {
    "title": "Location Service",
    "description": "This service is responsible for delivering messages to / from applications",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
...
    "/tickets": {
      "post": {
        "description": "Save Ticket",
        "operationId": "saveTicket",
        "tags": [
          "ticket"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ticket",
            "description": "Ticket",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Ticket"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "Ticket",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "400": {
            "description": "Bad Request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      },
      "get": {
        "description": "Page Tickets",
        "operationId": "page",
        "tags": [
          "ticket"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "query",
            "description": "Query String - default '{}'",
            "type": "string",
            "in": "query",
            "required": false
          },
          {
            "name": "limit",
            "description": "Limit - default 10",
            "in": "query",
            "required": false
          },
          {
            "name": "offset",
            "description": "Offset - default 0",
            "in": "query",
            "required": false
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "PageResult",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/PageResult"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    },
    "/tickets/{id}": {
      "put": {
        "description": "Update Ticket",
        "operationId": "updateTicket",
        "tags": [
          "ticket"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Ticket Id",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "ticket",
            "description": "Ticket",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Ticket"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "404": {
            "description": "Not Found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      },
      "$ref": "#/definitions/Error",
      "get": {
        "description": "Find Ticket By Id",
        "operationId": "findTicketById",
        "tags": [
          "ticket"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Ticket Id",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "404": {
            "description": "Not Found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    }
  },
  "definitions": {
...
    "PageResult": {
      "type": "object",
      "required": [
        "limit",
        "offset",
        "total"
      ],
      "properties": {
        "limit": {
          "type": "number"
        },
        "offset": {
          "type": "number"
        },
        "total": {
          "type": "number"
        }
      }
    },
    "Ticket": {
      "type": "object",
      "required": [
        "inflightAccountId",
        "ttl",
        "startTime",
        "destination"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "inflightAccountId": {
          "type": "string"
        },
        "startTime": {
          "type": "number",
          "format": "date"
        },
        "destination": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/Destination"
          }
        },
        "ttl": {
          "type": "integer"
        },
        "timestamp": {
          "type": "integer"
        }
      }
    },
...
    "Destination": {
      "type": "object",
      "required": [
        "street",
        "city",
        "state",
        "postalCode",
        "countryCode"
      ],
      "properties": {
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}